Move query_utils to /public (#983)
